move pinned items to own component
